refactor: move Heatmap example
