+++ examples/shared/src/examples/FillRasterLayer/Heatmap.tsx
